rename getServerSidePropsWrapper file
